requirements.txt
customtkinter==5.2.0
pillow==10.0.0
reportlab==4.0.4
PyInstaller==5.4.1